Exempt drawer navigation from throttling
Throttling was preventing the drawer from navigating to different screens, because the process dispatches actions at too high a rate. I also restored the navigation setup to its state before this branch, apart from adding the drawer example.
